+++ CHANGELOG.md
@@ -22,9 +22,11 @@ categories: project
 
 ---
 
-## 2021-03-31
+## 2021-04-02
 
-> 
+- 유저 메인 Controller 구현 
+  - 회원가입 이메일 인증번호 발송 및 확인 구현 `js, jsp, Action Controller`
+-